Make port assignments more flexible
The base port can be left unspecified and will then be chosen by the
system.  This is intended for use in Mtest, where a script can `from
tlstester import Certs, TLSTester` and then start TLSTester on a
background thread on an arbitrary port number and try to run, say,
mclient against it.

Also, if need be certain named ports can be forced to a certain port
number using the the 'preassigned' argument or the -a command line
option.

fixup flex